+++ packages/extension/src/view/devtools/pages/index.ts
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './antiCovertTracking';
+export * from './privacyProtection';
 export * from './privateAdvertising';
 export * from './siteBoundaries';
 export * from './learning';
