+++ packages/extension/src/view/devtools/pages/privacyProtection/antiCovertTracking.tsx
