+++ packages/extension/src/view/devtools/pages/privacyProtection/bounceTracking/index.tsx
