+++ packages/extension/src/view/devtools/pages/privacyProtection/index.ts
