+++ packages/extension/src/view/devtools/pages/privacyProtection/ipProtection/index.tsx
