+++ packages/extension/src/view/devtools/pages/privacyProtection/privateStateTokens/index.tsx
