+++ packages/extension/src/view/devtools/pages/privacyProtection/tests/index.tsx
