+++ packages/extension/src/view/devtools/pages/privacyProtection/userAgentReduction/index.tsx
